requirements.txt
asgiref==3.4.1
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
clickclick==20.10.2
connexion==2.12.0
coreapi==2.3.3
coreschema==0.0.4
dataclasses==0.8
Django==3.2.12
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
drf-yasg==1.20.0
Faker==4.0.0
Flask==2.0.3
greenlet==1.1.2
idna==3.3
importlib-metadata==4.8.3
inflection==0.5.1
iniconfig==1.1.1
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.3
jsonschema==3.2.0
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
mixer==7.0.6
mysqlclient==2.1.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
PyJWT==1.7.1
pyparsing==3.0.7
pyrsistent==0.18.0
pytest==7.0.1
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
SQLAlchemy==1.4.31
sqlparse==0.4.2
text-unidecode==1.3
tomli==1.2.3
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
zipp==3.6.0
